Bugfix: repair missing size units for PROBE tasks
If a team applies a workflow to a component, and subsequently change one
of the workflow steps to a PROBE task, the instantiated task may not
have any size metric stored on it. This was causing problems with
reverse sync operations. This change scans for and repairs that problem
on WBS Editor startup and save.