Bugfix: don't change planned time when marking overspent tasks complete
When a task is reassigned from an individual, or moved to a relaunched
project iteration, the HierarchySynchronizer will offer to close the
task in an individual's personal plan. As part of this change, the
estimated time is reduced to match the actual; that allows a relaunched
task to receive the remaining time, and cross-cycle rollups of estimated
time to remain unchanged. Unfortunately, this alteration was being
performed when the task was overspent as well. As a result,
underestimations were being swept under the rug.