Let coaches state that a project is being joined as another person
Sometimes, a coach may wish to join the same project several times from
a single personal dashboard. (This usage model can be used to support
what-if scenarios and other data explorations.) Add fields to the
project parameters page that allow a coach to enable this functionality,
and add logic to the Sync WBS to initialize default values.