Throw exception from WorkingDirectory.update if bundle mode has changed
If a working directory changes from bundled to unbundled while the app
is running, or if the target directory becomes unavailable, arrange for
the WorkingDirectory.update() method to throw an exception.

Among other things, this helps the WBS Editor fail fast on Save and
display a reasonable error message to the user.