Keep track of bundles that were replaced due to retention policies
When the retention granularity policy recommends that we replace an
earlier bundle, record the ID of that bundle in the replacement bundle.
This more complete bookkeeping will help with the future needs of merge
tracking and history reconstruction.

Note that if we replace a bundle many times, we're currently only
retaining the ID of the first bundle that was replaced. Any bundle ID
with the same name and device ID, and with a timestamp between the first
replacement and the final bundle, can be assumed to be one of the
replaced bundles.