Cache the manifests for bundles checked out in a working directory
A BundledWorkingDirectory keeps a list of HEAD refs for the bundles that
are locally checked out. This change arranges for the directory to keep
a local copy of the corresponding manifests as well. These will allow us
to remember what was checked out, even if the bundle gets deleted from
the official bundle directory by a retention/replacement operation.