When sorting tasks by label, respect value ordering from WBS columns
Custom WBS columns can offer a constrained list of values for the user
to choose from. When they do, the user enters the values in the custom
column editor in a particular order. For example, a Priority column
might have the values "high, medium, low". If the user sorts their Flat
View by that label, sort the labels into the same order they appeared in
the column spec.