Bugfix: code signing mismatch prevents launching with older profiles
In PD 2.6.6, I changed the certificate that was used to sign dashboard
code. This created an unexpected mismatch in security policy between the
new launcher and older launch profiles. As a result, once someone had
opened a dataset using 2.6.6 or higher, they would be unable to open any
dataset that uses a pre-2.6.6 version of the dashboard.