Configure socket connect/read timeouts to avoid startup/shutdown hang
The default socket timeouts in Java are very long (possibly infinite).
If a network, proxy server, or VPN becomes unresponsive, this can cause
our communications with the PDES to hang, resulting in an infinite wait
on startup or shutdown. This fix avoids that problem by setting
reasonable timeouts for socket connect/read.