Tweak syncDown age limits for update & validate of bridged import dir
In the past, BridgedImportDirectory.update would check with the server
if our last syncDown was over one second old; this changes the time
limit to 5 seconds. Similarly, the validate method would always check
with the server; now it skips the check if an update was completed in
the previous half a second.