Fix fork tracking logic to recognize parentage for replaced bundles
Bundle retention strategies mean a single bundle can get overwritten
multiple times, and the intermediate bundles can be discarded. When
another client is pointing to one of those "replaced" intermediate
bundles, it is pointing to a parent state of the final replacement
bundle. Recognize this relationship for fork-tracking and fast-forward
purposes.