In sync bundle mode, ensure newly created data files use unique names
When new nodes are created in the hierarchy, the DashHierarchy class
assigns filenames for new .dat/.def files. Historically, these have been
assigned sequentially. But when the directory is using sync bundle
storage, it's possible that two different computers can create new nodes
in parallel processes. Those new nodes must not choose the same names
for the files they create. This change arranges for new filenames to
include a suffix based on the device that created it.