Track the need for a projDump refresh after a WBS bundle merge
The WBS sync bundle merge process does not refresh the projDump.xml file
or the workflowDump.xml file. When performing a bundle merge, set a flag
to record this pending need. Perform a startup save if the flag is set
on startup, and clear the flag after performing each dump.