Move [Shared Folder] encoding resolution to FolderMappingManager
During a team project join operation, the setup wizard was detecting
situations where a "[Shared Folder]/relative/path" didn't map to a real
directory, and was searching for an existing directory that did match
(albeit with a slightly different relative path). This commit moves that
search logic into the FolderMappingManager class, to enable reuse.