Protect bundled "write single file" operations with a working dir lock
A previous change introduced a dedicated write lock for bundled working
directories to avoid overlapping syncUp/syncDown operations. This change
extends that lock protection to logic that is writing a single file,
such as a PDASH file or settings.xml. With this change, all writes to
the directory happen within the context of a lock.