Bugfix: repairCorruptFiles should not delete obsolete working heads
The original startup logic BundledWorkingDirectoryLocal included a step
to delete obsolete working heads (heads that claim to be checked out in
the local cached working dir, but which no longer appear in the bundle
dir). This was done to squash subsequent exceptions that might occur if
we attempt to perform operations against those nonexistent bundles.

That concern is no longer applicable, since the ManifestCachingHeadRefs
allow us to remember details about the bundles we checked out. And it
does the wrong thing if our locally checked out bundle happened to be
replaced. So this change removes that step.