Bugfix: don't flush leftover dirty files if sync bundle heads invalid
If the Team Dashboard crashes without flushing locally modified files to
a sync bundle directory, it detects this the next time it starts and
attempts to save the changes. But if the remote bundle dir has been
radically changed in the meantime (for example, by an end user who
moves/replaces entire data directories), that flush must be skipped.
This commit adds that check.