Arrange for early logging output to make it into log.txt
We cannot write to log.txt until we've locked the dashboard directory
for writing. Historically, this meant that log.txt didn't include any
logging output that preceded the acquisition of the lock. This change
arranges for all logging output to be captured in memory as early as
possible, and for that output to be copied to log.txt when it is opened.
As a result, information about network initialization and working
directory setup can be written into log.txt.