Use "sync needed" notifications to help users fix shared folder config
If the [Shared Folder] encoding of a project directory points to a
location that cannot be found, set the "sync needed" flag so users will
see a notification in their dashboard window. When they respond to that
notification, we will be allowed to deliver detailed info about the
config problem.

Reuse this new notification to bring problems to the users attention if
a cloud migration or export data operation suspects potentially bad
shared folder configuration.