Add logic to gather up the bundles for a month into a single pack
Bundles can be created every time data is saved, resulting in thousands
of files for a busy project. The large number of files can be a
liability for the sync client, either by slowing it down or by bumping
up against hardcoded limits.

This change introduces logic that can gather up all the bundles for a
particular month, and place them into a single XML/ZIP pair. It also
includes a utility to scan a directory and bundle all the months
preceding a particular time.