Never cache external resources that were extracted from a data backup
When a data backup is opened by the quick launcher, the external
resource directories will be extracted into a temporary location. We
never need to create CachingLocalImportDirectory objects for these
locations. Various mechanisms may prevent this from happening, but this
change adds "externalResources" to the never-cache setting as an extra
layer of checking.