Guard against partially uploaded bundles during fast forward/merge
When we publish a new bundle, that will involve a new XML and ZIP file,
along with changes to a HEADs file. But we have no guarantees that a
third-party sync client will upload all of those new files atomically.
If it has a large backlog of files that need to be uploaded (very
possible if someone has been offline for a while), it could potentially
publish the HEADs file before publishing the XML and ZIP.

When a different computer is performing a fast forward or merge, this
change detects that scenario during and ignores the partially published
bundle.