Maintain separate streams for stdout and stderr in ConsoleWindow
The ConsoleWindow installs handlers to intercept all output sent to
stderr and stdout, and copy them to a UI window and a log file. In the
past, both streams were merged into a single target stream, which wrote
to stdout after the processing above. This commit allows the streams to
remain separate, so data written to the Java System.err object will
eventually be written to stderr for the operating system process.