Bugfix: US-ASCII character encoding was causing PDES lock failures
When an individual opens a dataset on the PDES, they acquire a write
lock on the data. This lock includes various properties, including their
name and the name of the computer they are connecting from. If their
name includes non-ASCII characters, this was resulting in lock failures
down the line, due to an enhancement made in Process Dashboard 2.6.8.