Use a single set of WBS columns for all external systems
Previous versions of the WBS Editor would generate ID, Type, and Owner
columns for each external system. In practice, this led to a large
number of sparsely populated columns when a WBS was connected to
multiple systems. This change creates four columns for Source, ID, Type,
and Owner, and uses those to display read-only data for all systems.