Bugfix: "./data" mapping was shadowing some "./data/XXXX" mappings
A past change was made to allow team and personal projects to have a
project data directory of ".", which makes the project data area a
simple subdir of the dashboard data directory. This is a best practice,
and is now required for personal projects and for cloud-sync-based team
projects.

When you open a data backup from one of those dashboards, the external
resources will contain mappings for a number of "./data/XXXX" dirs.
Unfortunately, the previous enhancement was coded by adding a "./data"
mapping in with those other entries, and the nondeterminism of HashMap
made it appear in a random place in the list. When a quick-launched
dashboard was resolving paths, it would find a match on "./data" and not
look any further to find the more precise match. This was resulting in
"missing data" problems for half the projects in the dashboard, even
though the files were present in the backup.