Assign new IDs to more incoming added WBS nodes when merging
When a 3-way merge is performed on the WBS, and nodes have been added in
the incoming branch, those nodes are checked and possibly assigned new
unique IDs. Historically, new IDs were only assigned if the incoming IDs
collided with an ID currently in use by the main branch. With this
change, a new ID is assigned if the incoming ID falls anywhere within
the range that has historically been used by the main branch. As a
result, the incoming branch will never reuse an ID that might have been
used by the main branch in the past.