When merging WBS bundles, write node ID changes into bundle metadata
If two people simultaneously perform offline edits to the WBS in a
bundled directory, and both add new nodes to the WBS, the unique IDs of
those added nodes may collide. The 3-way merge logic modifies the IDs in
the incoming branch to fix these collisions. When this occurs during a
bundle merge, write the ID changes into a metadata file in the bundle.