Bugfix: apply size metric ID changes into applied workflow PROBE tasks
When a workflow contains a PROBE step, that step specifies the custom
size metric that it uses for estimating.  When this workflow is applied
to the WBS, this custom size metric ID is copied into the task that is
created.  If that size metric ID changes later as the result of a merge
operation, this change must be propagated into the WBS nodes that were
created for the PROBE step in question.