Perform reverse sync selectively when PDASH files predate a merge
When a bundle merge occurs, the WBS node IDs in the incoming branch may
be modified to avoid ID collsions. However, PDASH files in the directory
that were written before the merge may contain reverse-sync data that
refers to those obsolete incoming IDs. Those instructions should be
ignored; otherwise, they would incorrectly be applied to the main nodes
that the IDs collided with.

This change introduces a WBSSyncFilter that can detect a pre-merge PDASH
file; identify the ancestor bundle for the subsequent 3-way merge; and
ignore reverse sync data for nodes that were added by the incoming
branch.