Clean up extraneous peg files that are left over in a heads directory
A single peg file is needed to indicate the active head ref for a PDASH
file. But it's possible that multiple peg files will end up in a heads
directory. Only the newest one matters; the older ones are ignored. This
change detects these older files and cleans them up, including files
that might have been created by a sync client in response to an imagined
editing conflict.