Safeguard orphaned WBS tasks from being incorrectly reused
If a task has actual data when it is deleted from the WBS, the sync
operation will offer to mark it complete instead of deleting it. If a
later WBS edit creates a new unrelated node with the same WBS ID, make
sure we don't resurrect and repurpose the orphaned task for it.